backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
openai>=1.3.5
requests==2.31.0
flask-cors==4.0.0
gunicorn==21.2.0
google-generativeai>=0.3.2
grpcio>=1.74.0
anyio>=4.8.0,<5.0.0
pydantic>=2.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
alembic==1.13.1
pytest>=7.0.0
psycopg2-binary>=2.9.0
